LPS-123218 Move exceptions to package
