+++ modules/dxp/apps/saml/saml-web/src/main/java/com/liferay/saml/web/internal/exception/UserIdentifierExpressionException.java
@@ -12,7 +12,7 @@
  *
  */
 
-package com.liferay.saml.web.internal;
+package com.liferay.saml.web.internal.exception;
 
 import com.liferay.portal.kernel.exception.PortalException;
