+++ modules/dxp/apps/saml/saml-web/src/main/java/com/liferay/saml/web/internal/opensaml/integration/field/expression/handler/AttributeMappingSamlSpIdpConnectionFieldExpressionHandler.java
@@ -23,8 +23,8 @@ import com.liferay.portal.kernel.util.Validator;
 import com.liferay.saml.opensaml.integration.field.expression.handler.SamlSpIdpConnectionFieldExpressionHandler;
 import com.liferay.saml.opensaml.integration.processor.context.SamlSpIdpConnectionProcessorContext;
 import com.liferay.saml.persistence.model.SamlSpIdpConnection;
-import com.liferay.saml.web.internal.UserAttributeMappingException;
-import com.liferay.saml.web.internal.UserIdentifierExpressionException;
+import com.liferay.saml.web.internal.exception.UserAttributeMappingException;
+import com.liferay.saml.web.internal.exception.UserIdentifierExpressionException;
 
 import java.io.StringWriter;
